pyproject.toml
[project]
name = "chatbot-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "openai-agents>=0.0.13",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]
